+++ BlogPostify.Service/Interfaces/Users/IUserService.cs
@@ -5,9 +5,9 @@ namespace BlogPostify.Service.Interfaces.Users;
 
 public interface IUserService
 {
-    Task<bool> RemoveAsync(long id);
-    Task<UserForResultDto> RetrieveByIdasync(long id);
+    Task<bool> RemoveAsync(int id);
+    Task<UserForResultDto> RetrieveByIdasync(int id);
     Task<UserForResultDto> AddAsync(UserForCreationDto dto);
-    Task<UserForResultDto> ModifyAsync(long id, UserForUpdateDto dto);
+    Task<UserForResultDto> ModifyAsync(int id, UserForUpdateDto dto);
     Task<IEnumerable<UserForResultDto>> RetrieveAllAsync(PaginationParams @params);
 }
